Rename reducer.ts to state.reducer.ts
