+++ visualization/app/codeCharta/state/store/state.reducer.ts
